Run only if there any program to run

+++ z.go
@@ -35,41 +35,43 @@ func (z *Goz) Start() {
 	}
 
 	log.Println(args.args)
-
-	for _, p := range args.programs {
-		z.Add(NewGod(p, args.args))
-	}
-	// need to handle panic and shut down others
-	for _, d := range z.gods {
-		d.Start()
-	}
-	sigc := make(chan os.Signal, 1)
-	signal.Notify(sigc, os.Kill, os.Interrupt, syscall.SIGTERM, syscall.SIGHUP)
-	for {
-		sig := <-sigc
-		log.Printf("Got signal %v", sig)
-		switch sig {
-		case syscall.SIGHUP:
-			z.Restart()
-			break
-		case syscall.SIGTERM:
-			z.Stop()
-			goto programExit
-		case os.Kill:
-			z.Stop()
-			goto programExit
-		case os.Interrupt:
-			z.Stop()
-			goto programExit
-		default:
-			log.Printf("Unhandled signal %v, stop program", sig)
-			z.Stop()
-			goto programExit
-
+	
+	if len(args.programs) > 0 {
+		for _, p := range args.programs {
+			z.Add(NewGod(p, args.args))
 		}
+		// need to handle panic and shut down others
+		for _, d := range z.gods {
+			d.Start()
+		}
+		sigc := make(chan os.Signal, 1)
+		signal.Notify(sigc, os.Kill, os.Interrupt, syscall.SIGTERM, syscall.SIGHUP)
+		for {
+			sig := <-sigc
+			log.Printf("Got signal %v", sig)
+			switch sig {
+			case syscall.SIGHUP:
+				z.Restart()
+				break
+			case syscall.SIGTERM:
+				z.Stop()
+				goto programExit
+			case os.Kill:
+				z.Stop()
+				goto programExit
+			case os.Interrupt:
+				z.Stop()
+				goto programExit
+			default:
+				log.Printf("Unhandled signal %v, stop program", sig)
+				z.Stop()
+				goto programExit
+
+			}
+		}
+	programExit:
+		log.Println("Program exit")
 	}
-programExit:
-	log.Println("Program exit")
 }
 
 func (z *Goz) Stop() {
